Maintain: Standardized the code style of string.

+++ include/string.h
@@ -48,7 +48,7 @@ extern char *strim(char *s);
  * Memory manipulation
  */
 extern void *memset(void *s, int c, size_t n);
-extern void *memcpy(void *dest, const void*src, size_t len);
+extern void *memcpy(void *dest, const void *src, size_t len);
 extern void *memmove(void *dest, const void *src, size_t n);
 extern void *memchr(const void *s, int c, size_t n);
 extern void *memscan(void *addr, int c, size_t size);
@@ -59,8 +59,8 @@ extern int memcmp(const void *s1, const void *s2, size_t n);
  */
 extern int strcount(const char *str, char c);
 extern int strncount(const char *str, char c, size_t size);
-extern int memcount(const void *addr, long c, size_t size);
-extern void *memdiff(const void *addr, long c, size_t size);
+extern int memcount(const void *addr, int c, size_t size);
+extern void *memdiff(const void *addr, int c, size_t size);
 extern char *skip_spaces(const char *path);
 extern char *basename(const char *path);
